requirements.txt
antiorm==1.2.1
anyio==4.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cryptography==41.0.6
dask==2023.11.0
db==0.1.1
db-sqlite3==0.0.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fsspec==2023.10.0
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-metadata==6.8.0
locket==1.0.0
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
pandas==2.0.3
partd==1.4.1
pycparser==2.21
PyMySQL==1.0.2
pyTelegramBotAPI==4.14.0
python-dateutil==2.8.2
python-telegram-bot==20.7
pytz==2023.3.post1
PyYAML==6.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
telebot==0.0.5
telegram==0.0.1
toolz==0.12.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.4
xlrd==2.0.1
xlsx2csv==0.8.2
Yaml8==0.1.2
zipfile36==0.1.3
zipp==3.17.0
fastparquet==0.7.0
fileinput == 
tempfile ==
shutil ==
